feat: Add font theme
